Moved redundant file names to __init__.py files.
