+++ ScoutSuite/providers/aws/resources/cloudformation/__init__.py
@@ -2,7 +2,7 @@ import re
 
 from ScoutSuite.providers.aws.resources.regions import Regions
 from ScoutSuite.providers.aws.resources.resources import AWSResources
-from ScoutSuite.providers.aws.facade.facade import AWSFacade
+from ScoutSuite.providers.aws.facade import AWSFacade
 
 
 class Stacks(AWSResources):
