+++ ScoutSuite/providers/aws/resources/cloudtrail/__init__.py
@@ -1,6 +1,6 @@
 import time
 
-from ScoutSuite.providers.aws.facade.facade import AWSFacade
+from ScoutSuite.providers.aws.facade import AWSFacade
 from ScoutSuite.providers.aws.resources.regions import Regions
 from ScoutSuite.providers.aws.resources.resources import AWSResources
 from ScoutSuite.providers.utils import get_non_provider_id
