+++ ScoutSuite/providers/aws/resources/cloudwatch/__init__.py
@@ -2,7 +2,7 @@ from ScoutSuite.providers.aws.resources.regions import Regions
 from ScoutSuite.providers.aws.resources.resources import AWSResources
 
 from ScoutSuite.providers.utils import get_non_provider_id
-from ScoutSuite.providers.aws.facade.facade import AWSFacade
+from ScoutSuite.providers.aws.facade import AWSFacade
 
 
 class Alarms(AWSResources):
