+++ ScoutSuite/providers/aws/resources/ec2/__init__.py
