+++ ScoutSuite/providers/aws/resources/sqs/__init__.py
@@ -1,6 +1,6 @@
 import json
 
-from ScoutSuite.providers.aws.facade.facade import AWSFacade
+from ScoutSuite.providers.aws.facade import AWSFacade
 from ScoutSuite.providers.aws.resources.regions import Regions
 from ScoutSuite.providers.aws.resources.resources import AWSResources
