+++ ScoutSuite/providers/aws/services.py
@@ -1,32 +1,32 @@
 # -*- coding: utf-8 -*-
 
-from ScoutSuite.providers.aws.resources.awslambda.service import Lambdas
-from ScoutSuite.providers.aws.resources.cloudwatch.service import CloudWatch
-from ScoutSuite.providers.aws.resources.cloudformation.service import CloudFormation
-from ScoutSuite.providers.aws.resources.cloudtrail.service import CloudTrail
-from ScoutSuite.providers.aws.resources.directconnect.service import DirectConnect
-from ScoutSuite.providers.aws.resources.ec2.service import EC2
-from ScoutSuite.providers.aws.resources.efs.service import EFS
-from ScoutSuite.providers.aws.resources.elasticache.service import ElastiCache
-from ScoutSuite.providers.aws.resources.elb.service import ELB
-from ScoutSuite.providers.aws.resources.elbv2.service import ELBv2
-from ScoutSuite.providers.aws.resources.iam.service import IAM
-from ScoutSuite.providers.aws.resources.emr.service import EMR
-from ScoutSuite.providers.aws.resources.route53.service import Route53
-from ScoutSuite.providers.aws.resources.rds.service import RDS
-from ScoutSuite.providers.aws.resources.redshift.service import Redshift
-from ScoutSuite.providers.aws.resources.s3.service import S3
-from ScoutSuite.providers.aws.resources.vpc.service import VPC
-from ScoutSuite.providers.aws.resources.sqs.service import SQS
-from ScoutSuite.providers.aws.resources.ses.service import SES
-from ScoutSuite.providers.aws.resources.sns.service import SNS
+from ScoutSuite.providers.aws.resources.awslambda import Lambdas
+from ScoutSuite.providers.aws.resources.cloudwatch import CloudWatch
+from ScoutSuite.providers.aws.resources.cloudformation import CloudFormation
+from ScoutSuite.providers.aws.resources.cloudtrail import CloudTrail
+from ScoutSuite.providers.aws.resources.directconnect import DirectConnect
+from ScoutSuite.providers.aws.resources.ec2 import EC2
+from ScoutSuite.providers.aws.resources.efs import EFS
+from ScoutSuite.providers.aws.resources.elasticache import ElastiCache
+from ScoutSuite.providers.aws.resources.elb import ELB
+from ScoutSuite.providers.aws.resources.elbv2 import ELBv2
+from ScoutSuite.providers.aws.resources.iam import IAM
+from ScoutSuite.providers.aws.resources.emr import EMR
+from ScoutSuite.providers.aws.resources.route53 import Route53
+from ScoutSuite.providers.aws.resources.rds import RDS
+from ScoutSuite.providers.aws.resources.redshift import Redshift
+from ScoutSuite.providers.aws.resources.s3 import S3
+from ScoutSuite.providers.aws.resources.vpc import VPC
+from ScoutSuite.providers.aws.resources.sqs import SQS
+from ScoutSuite.providers.aws.resources.ses import SES
+from ScoutSuite.providers.aws.resources.sns import SNS
 from ScoutSuite.providers.base.configs.services import BaseServicesConfig
-from ScoutSuite.providers.aws.facade.facade import AWSFacade
+from ScoutSuite.providers.aws.facade import AWSFacade
 
 try:
-    from ScoutSuite.providers.aws.resources.dynamodb.service_private import DynamoDB
-    from ScoutSuite.providers.aws.resources.config.service_private import Config
-    from ScoutSuite.providers.aws.resources.kms.service_private import KMS
+    from ScoutSuite.providers.aws.resources.private_dynamodb import DynamoDB
+    from ScoutSuite.providers.aws.resources.private_config import Config
+    from ScoutSuite.providers.aws.resources.private_kms import KMS
 except ImportError:
     DynamoDB = None
     Config = None
