+++ ScoutSuite/providers/azure/facade/__init__.py
