+++ ScoutSuite/providers/azure/resources/keyvault/__init__.py
@@ -1,4 +1,4 @@
-from ScoutSuite.providers.azure.facade.facade import AzureFacade
+from ScoutSuite.providers.azure.facade import AzureFacade
 from ScoutSuite.providers.base.configs.resources import Resources
 from ScoutSuite.providers.utils import get_non_provider_id
