+++ ScoutSuite/providers/azure/resources/network/__init__.py
