+++ ScoutSuite/providers/azure/resources/securitycenter/__init__.py
