+++ ScoutSuite/providers/azure/resources/securitycenter/pricings.py
@@ -1,4 +1,4 @@
-from ScoutSuite.providers.azure.facade.facade import AzureFacade
+from ScoutSuite.providers.azure.facade import AzureFacade
 from ScoutSuite.providers.base.configs.resources import Resources
