+++ ScoutSuite/providers/azure/resources/sqldatabase/replication_links.py
@@ -1,5 +1,5 @@
 from ScoutSuite.providers.base.configs.resources import Resources
-from ScoutSuite.providers.azure.facade.facade import AzureFacade
+from ScoutSuite.providers.azure.facade import AzureFacade
 
 
 class ReplicationLinks(Resources):
