+++ ScoutSuite/providers/azure/resources/sqldatabase/server_azure_ad_administrators.py
@@ -1,6 +1,6 @@
 from msrestazure.azure_exceptions import CloudError
 
-from ScoutSuite.providers.azure.facade.facade import AzureFacade
+from ScoutSuite.providers.azure.facade import AzureFacade
 from ScoutSuite.providers.base.configs.resources import Resources
