+++ ScoutSuite/providers/azure/resources/storageaccounts/__init__.py
