+++ ScoutSuite/providers/azure/resources/storageaccounts/blob_containers.py
@@ -1,5 +1,5 @@
 from ScoutSuite.providers.base.configs.resources import Resources
-from ScoutSuite.providers.azure.facade.facade import AzureFacade
+from ScoutSuite.providers.azure.facade import AzureFacade
 
 
 class BlobContainers(Resources):
