+++ ScoutSuite/providers/azure/services.py
@@ -1,28 +1,28 @@
 # -*- coding: utf-8 -*-
 
 from ScoutSuite.providers.azure.authentication_strategy import AzureCredentials
-from ScoutSuite.providers.azure.facade.facade import AzureFacade
-from ScoutSuite.providers.azure.resources.keyvault.key_vaults import KeyVaults
-from ScoutSuite.providers.azure.resources.network.networks import Networks
-from ScoutSuite.providers.azure.resources.securitycenter.security_center import SecurityCenter
-from ScoutSuite.providers.azure.resources.sqldatabase.servers import Servers
-from ScoutSuite.providers.azure.resources.storageaccounts.storageaccounts import StorageAccounts
+from ScoutSuite.providers.azure.facade import AzureFacade
+from ScoutSuite.providers.azure.resources.keyvault import KeyVaults
+from ScoutSuite.providers.azure.resources.network import Networks
+from ScoutSuite.providers.azure.resources.securitycenter import SecurityCenter
+from ScoutSuite.providers.azure.resources.sqldatabase import Servers
+from ScoutSuite.providers.azure.resources.storageaccounts import StorageAccounts
 from ScoutSuite.providers.base.configs.services import BaseServicesConfig
 
 try:
-    from ScoutSuite.providers.azure.resources.appgateway_private.application_gateways_private import ApplicationGateways
+    from ScoutSuite.providers.azure.resources.private_appgateway import ApplicationGateways
 except ImportError:
     pass
 try:
-    from ScoutSuite.providers.azure.resources.rediscache_private.redis_caches_private import RedisCaches
+    from ScoutSuite.providers.azure.resources.private_rediscache import RedisCaches
 except ImportError:
     pass
 try:
-    from ScoutSuite.providers.azure.resources.appservice_private.web_applications_private import WebApplications
+    from ScoutSuite.providers.azure.resources.appservice_private import WebApplications
 except ImportError:
     pass
 try:
-    from ScoutSuite.providers.azure.resources.loadbalancer_private.load_balancers_private import LoadBalancers
+    from ScoutSuite.providers.azure.resources.private_loadbalancer import LoadBalancers
 except ImportError:
     pass
